% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prob_living_2014}
\alias{prob_living_2014}
\title{Probability of Living in 2014}
\format{A data frame with three variables:
\describe{
\item{\code{year}}{year born}
\item{\code{sex}}{Sex}
\item{\code{px}}{Probability that someone born in year x is alive as of Jan 1, 2014}
}}
\usage{
prob_living_2014
}
\description{
The probabily that someone was born in a given year is alive today.
Derived from the extended life tables
}
\keyword{datasets}

